+++ RELEASENOTES.md
@@ -7,6 +7,8 @@
         results in a call to `Player.Listener#onTimelineChanged` with
         `reason=Player.TIMELINE_CHANGE_REASON_PLAYLIST_CHANGED`
         ([#9889](https://github.com/google/ExoPlayer/issues/9889)).
+    *   For progressive media, only include selected tracks in buffered position
+        ([#10361](https://github.com/google/ExoPlayer/issues/10361)).
 *   Extractors:
     *   Add support for AVI
         ([#2092](https://github.com/google/ExoPlayer/issues/2092)).